#chore(docker): created a docker config for intelliJ based jekyll development
So I followed along a great video on locally developing a Jekyll based GitHub page on YouTube and struggled quite a bit with the setup. First time with docker and all. Tutorials are mostly based out of VSCode. For my purposes I added in an open-jdk install. This is so that intelliJ can work from my containerized env.